Graphs AND Maps
A bunch of progress today. The website no longer looks really really
bad, just kind of bad. I got some graphs working using this Graphjs
library. I'm not sure if that's what we want to use in the end but for
now it looks nice. Also I got some Google maps in there that are
centered at the latitude and longitude. Later there will be markers and
such. Done for tonight though...