Continued Developing Dynamic index
Continued to develop Dynamix Index

It now it more interactive and supports Temperature, Radiation, and CO2
graphs. Still working on getting a map thrown in there as well.

Also created a cron job that inserts random values into the database
every hour, this should give a good set of data